+++ scoop-viewer.py
@@ -27,7 +27,7 @@ class Viewer(wx.Frame): # pylint: disable=too-many-ancestors
 
         sizer = wx.BoxSizer(wx.VERTICAL)
         self.browser = wx.html2.WebView.New(self)
-        #self.browser.EnableContextMenu(False)
+        self.browser.EnableContextMenu(False)
         self.browser.Bind(wx.html2.EVT_WEBVIEW_NEWWINDOW, Viewer.handle_new_window)
 
         sizer.Add(self.browser, 1, wx.EXPAND, 10)