+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### Unreleased
+- Added settings page.
+- Allow showing installed packages at the top of the bucket list.
+
 ## [v0.8] - 2019-06-13
 ### Added 
 - Make it possible to add new bucket.