+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Unreleased
 - Added settings page.
 - Allow showing installed packages at the top of the bucket list.
+- Added bottom panel with the number of packages and an option to show only installed.
 
 ## [v0.8] - 2019-06-13
 ### Added